requirements.txt
Flask==2.0.1
pymongo==3.11.4
gridfs==1.1.0
docxtpl==0.11.4
python-docx==0.8.11
docx2pdf==0.1.7
mammoth==1.4.2
bcrypt==3.2.0
